+++ src/IIIFAuth2/IIIFAuth2.API.Tests/Integration/ServicesTests.cs
@@ -114,11 +114,11 @@ public class ServicesTests : IClassFixture<AuthWebApplicationFactory>
             {
                 new AuthAccessService2
                 {
-                    Id = "http://localhost/auth/v2/99/clickthrough",
+                    Id = "http://localhost/auth/v2/access/99/clickthrough",
                     Profile = "active",
                     Service = new List<IService>
                     {
-                        new AuthAccessTokenService2 { Id = "http://localhost/auth/v2/99/token", },
+                        new AuthAccessTokenService2 { Id = "http://localhost/auth/v2/access/99/token", },
                         new AuthLogoutService2
                         {
                             Id = "http://localhost/auth/v2/99/clickthrough/logout",
@@ -152,11 +152,11 @@ public class ServicesTests : IClassFixture<AuthWebApplicationFactory>
             {
                 new AuthAccessService2
                 {
-                    Id = "http://localhost/auth/v2/99/clickthrough",
+                    Id = "http://localhost/auth/v2/access/99/clickthrough",
                     Profile = "active",
                     Service = new List<IService>
                     {
-                        new AuthAccessTokenService2 { Id = "http://localhost/auth/v2/99/token", },
+                        new AuthAccessTokenService2 { Id = "http://localhost/auth/v2/access/99/token", },
                         new AuthLogoutService2
                         {
                             Id = "http://localhost/auth/v2/99/clickthrough/logout",
