+++ src/IIIFAuth2/IIIFAuth2.API/Features/Access/Views/CloseWindow.cshtml
